@@ -1,15 +1,21 @@
 module github.com/rprtr258/gena
 
-go 1.21
+go 1.22rc1
 
 require (
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0
+	github.com/schollz/progressbar/v3 v3.14.1
 	github.com/stretchr/testify v1.7.0
+	golang.org/x/exp v0.0.0-20231219180239-dc181d75b848
 	golang.org/x/image v0.14.0
 )
 
 require (
-	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rivo/uniseg v0.4.4 // indirect
+	golang.org/x/sys v0.14.0 // indirect
+	golang.org/x/term v0.14.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
