@@ -1,6 +1,7 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math"
 )
@@ -10,7 +11,7 @@ import (
 //   - density: Control the density of the circle.
 //   - circleGap: Identify the gap between two circles.
 func BlackHole(
-	c Canvas,
+	c *image.RGBA,
 	lineWidth float64,
 	lineColor color.RGBA,
 	circleN int,
@@ -19,17 +20,17 @@ func BlackHole(
 	noise PerlinNoise,
 	alpha float64, // [0, 1)
 ) {
-	dc := NewContextForRGBA(c.Img())
+	dc := NewContextForRGBA(c)
 	kMax := Remap(alpha, 0, 1, 0.5, 1)
 	dc.SetLineWidth(0.4)
 	dc.SetColor(lineColor)
 
-	for i := 0; i < circleN; i++ {
-		radius := float64(c.Width/10) + float64(i)*0.05
+	for i := range circleN {
+		radius := float64(c.Bounds().Dx()/10) + float64(i)*0.05
 		k := kMax * math.Sqrt(float64(i)/float64(circleN))
 		noisiness := density * math.Pow(float64(i)/float64(circleN), 2)
 
-		base := c.Size() / 2
+		base := Size(c) / 2
 		for theta := 0.0; theta < 2*math.Pi; theta += 2 * math.Pi / 360 {
 			r1 := math.Cos(theta) + 1
 			r2 := math.Sin(theta) + 1
