@@ -1,6 +1,7 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math"
 	"math/rand"
@@ -10,41 +11,41 @@ type circleGrid struct {
 	circleNumMin, circleNumMax int
 }
 
-func (cg *circleGrid) grid(dc *Context, c Canvas) {
+func (cg *circleGrid) grid(dc *Context, c *image.RGBA) {
 	var segment int = 100
-	w := float64(c.Width) / float64(segment)
+	w := float64(c.Bounds().Dx()) / float64(segment)
 
 	dc.SetColor(color.RGBA{255, 255, 255, 255})
 	dc.SetLineWidth(0.6)
-	for i := 0; i < segment; i++ {
-		dc.DrawLine(complex(0, float64(i)*w), complex(float64(c.Width), float64(i)*w))
+	for i := range segment {
+		dc.DrawLine(complex(0, float64(i)*w), complex(float64(c.Bounds().Dx()), float64(i)*w))
 		dc.Stroke()
 	}
 
-	for j := 0; j < segment; j++ {
-		dc.DrawLine(complex(float64(j)*w, 0), complex(float64(j)*w, float64(c.Height)))
+	for j := range segment {
+		dc.DrawLine(complex(float64(j)*w, 0), complex(float64(j)*w, float64(c.Bounds().Dy())))
 		dc.Stroke()
 	}
 }
 
 // Generative draws a circle grid image.
-func CircleGrid(c Canvas, colorSchema []color.RGBA, lineWidth float64, circleNumMin, circleNumMax int) {
+func CircleGrid(c *image.RGBA, colorSchema []color.RGBA, lineWidth float64, circleNumMin, circleNumMax int) {
 	cg := &circleGrid{
 		circleNumMin: circleNumMin,
 		circleNumMax: circleNumMax,
 	}
 
-	dc := NewContextForRGBA(c.Img())
+	dc := NewContextForRGBA(c)
 	cg.grid(dc, c)
-	dc.Translate(c.Size() / 2)
+	dc.Translate(Size(c) / 2)
 	dc.Scale(complex(0.9, 0.9))
-	dc.Translate(-c.Size() / 2)
+	dc.Translate(-Size(c) / 2)
 
 	seg := RandomRangeInt(cg.circleNumMin, cg.circleNumMax)
-	w := float64(c.Width) / float64(seg)
+	w := float64(c.Bounds().Dx()) / float64(seg)
 
-	for i := 0; i < seg; i++ {
-		for j := 0; j < seg; j++ {
+	for i := range seg {
+		for j := range seg {
 			v := Plus(Mul(complex(float64(i), float64(j)), w), w/2)
 			dc.SetColor(colorSchema[rand.Intn(len(colorSchema))])
 			dc.DrawCircleV2(v, w/2*RandomFloat64(0.1, 0.5))
@@ -68,7 +69,7 @@ func CircleGrid(c Canvas, colorSchema []color.RGBA, lineWidth float64, circleNum
 					n := RandomRangeInt(1, 4) * 2
 					dc.DrawCircleV2(0, r)
 					dc.Stroke()
-					for i := 0; i < n; i++ {
+					for range n {
 						dc.Rotate(math.Pi * 2 / float64(n))
 						dc.DrawCircleV2(complex(r, 0), r*0.1)
 						dc.Fill()
@@ -76,7 +77,7 @@ func CircleGrid(c Canvas, colorSchema []color.RGBA, lineWidth float64, circleNum
 				case 2:
 					n := RandomRangeInt(8, 20)
 					theta := math.Pi * 0.5 * float64(RandomRangeInt(1, 5))
-					for i := 0; i < n; i++ {
+					for i := range n {
 						d := float64(i) / float64(n)
 						if d > r*0.1 {
 							d = r * 0.1
@@ -87,7 +88,7 @@ func CircleGrid(c Canvas, colorSchema []color.RGBA, lineWidth float64, circleNum
 					}
 				case 3:
 					n := RandomRangeInt(5, 20)
-					for i := 0; i < n; i++ {
+					for range n {
 						dc.Rotate(math.Pi * 2 / float64(n))
 						dc.DrawLine(complex(r/2, 0), complex(r*2/3-r*0.05, 0))
 						dc.Stroke()
