@@ -1,6 +1,7 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math"
 	"math/rand"
@@ -13,17 +14,17 @@ type circleLoop2 struct {
 
 // CircleLoop2 draws a circle composed by colored circles.
 //   - depth: Control the number of circles.
-func CircleLoop2(c Canvas, colorSchema []color.RGBA, depth int) {
-	ctex := NewContextForRGBA(c.Img())
-	ctex.Translate(c.Size() / 2)
+func CircleLoop2(c *image.RGBA, colorSchema []color.RGBA, depth int) {
+	dc := NewContextForRGBA(c)
+	dc.Translate(Size(c) / 2)
 	cl := &circleLoop2{
 		noise:       NewPerlinNoiseDeprecated(),
 		colorSchema: colorSchema,
 	}
-	cl.recursionDraw(ctex, c, float64(c.Width), depth)
+	cl.recursionDraw(dc, c, float64(c.Bounds().Dx()), depth)
 }
 
-func (cl *circleLoop2) recursionDraw(dc *Context, c Canvas, x float64, depth int) {
+func (cl *circleLoop2) recursionDraw(dc *Context, c *image.RGBA, x float64, depth int) {
 	if depth <= 0 {
 		return
 	}
@@ -40,8 +41,8 @@ func (cl *circleLoop2) recursionDraw(dc *Context, c Canvas, x float64, depth int
 	noise = math.Pow(noise, 0.5)
 	a2 := Remap(noise, 0.15, 0.85, 0.1, 0.6)
 
-	px := float64(c.Height) * math.Pow(x/float64(c.Height), a2)
-	py := float64(c.Height) * (math.Pow(1-x/float64(c.Height), a2) -
+	px := float64(c.Bounds().Dy()) * math.Pow(x/float64(c.Bounds().Dy()), a2)
+	py := float64(c.Bounds().Dy()) * (math.Pow(1-x/float64(c.Bounds().Dy()), a2) -
 		RandomFloat64(0, RandomFloat64(0.18, RandomFloat64(0.18, 0.7))))
 
 	dc.SetColor(cl.colorSchema[rand.Intn(len(cl.colorSchema))])
@@ -54,18 +55,18 @@ func (cl *circleLoop2) recursionDraw(dc *Context, c Canvas, x float64, depth int
 	r := math.Pow(rand.Float64(), 2) * 50
 
 	if rand.Float64() < 0.7 {
-		for i := 0; i < nCircles; i++ {
+		for i := range nCircles {
 			dc.DrawCircleV2(complex(px, py)*0.39, rand.Float64()*float64(i)*r/float64(nCircles))
 			dc.Stroke()
 		}
 	} else {
-		for i := 0; i < nCircles; i++ {
+		for i := range nCircles {
 			dc.DrawCircleV2(complex(px, py)*0.39, float64(i)*r/float64(nCircles))
 			dc.Stroke()
 		}
 	}
 
-	dc.Rotate(x / float64(c.Height) * 0.2)
+	dc.Rotate(x / float64(c.Bounds().Dy()) * 0.2)
 
 	cl.recursionDraw(dc, c, 1*x/4.0, depth-1)
 	cl.recursionDraw(dc, c, 2*x/4.0, depth-1)
