@@ -1,6 +1,7 @@
 package gena
 
 import (
+	"image"
 	"math"
 )
 
@@ -11,49 +12,55 @@ type dot struct {
 }
 
 // CircleNoise draws a circle with perlin noise.
-//   - dotsN: The number of dot.
-//   - colorMin: The minimum color.
-//   - colorMax: The maximum color.
-func CircleNoise(c Canvas, lineWidth float64, aalpha uint8, dotsN int, colorMin, colorMax float64, iters int) {
-	ctex := NewContextForRGBA(c.Img())
-	ctex.SetLineWidth(2.0)
-	ctex.SetColor(Black)
-	radius := float64(c.Width) * 0.8 / 2
-	ctex.DrawCircleV2(c.Size()/2, radius)
-	ctex.Stroke()
+//   - angles: The number of dot of random float64-s
+//   - colorMin: minimum color
+//   - colorMax: maximum color
+func CircleNoise(
+	c *image.RGBA,
+	lineWidth float64,
+	aalpha uint8,
+	angles []float64,
+	colorMin, colorMax float64,
+	iters int,
+	noise PerlinNoise,
+) {
+	dc := NewContextForRGBA(c)
+	dc.SetLineWidth(2.0)
+	dc.SetColor(Black)
+	radius := float64(c.Bounds().Dx()) * 0.8 / 2
+	dc.DrawCircleV2(Size(c)/2, radius)
+	dc.Stroke()
 	// ctex.Clip()
 
-	dots := make([]dot, dotsN)
-	for i := 0; i < dotsN; i++ {
-		theta := RandomFloat64(0, math.Pi*2)
-		v := c.Size()/2 + Polar(radius, theta)
+	dots := make([]dot, len(angles))
+	for i, angle := range angles {
+		v := Size(c)/2 + Polar(radius, angle*math.Pi*2)
 		dots[i] = dot{pos: v, prev: v, count: 0}
 	}
 
 	const factor = 0.008
-	noise := NewPerlinNoiseDeprecated()
-	for j := 0; j < iters; j++ {
+	for range iters {
 		for i := range dots {
 			n := noise.NoiseV2(dots[i].pos * factor)
 			alpha := math.Pi * (n*2 + float64(dots[i].count))
 			nn := Polar(2, alpha)
 			dots[i].prev, dots[i].pos = dots[i].pos, dots[i].pos+nn
-			if Dist(c.Size()/2, dots[i].pos) > radius+2 {
+			if Dist(Size(c)/2, dots[i].pos) > radius+2 {
 				dots[i].count += 1
 			}
 
-			if Dist(c.Size()/2, dots[i].pos) < radius &&
-				Dist(c.Size()/2, dots[i].prev) < radius {
-				ctex.SetLineWidth(lineWidth)
+			if Dist(Size(c)/2, dots[i].pos) < radius &&
+				Dist(Size(c)/2, dots[i].prev) < radius {
+				dc.SetLineWidth(lineWidth)
 				rgb := HSV{
 					H: int(Remap(n, 0, 1, colorMin, colorMax)),
 					S: 100,
 					V: 20,
 				}.ToRGB(100, 100, 100)
 				rgb.A = aalpha
-				ctex.SetColor(rgb)
-				ctex.DrawLine(dots[i].prev, dots[i].pos)
-				ctex.Stroke()
+				dc.SetColor(rgb)
+				dc.DrawLine(dots[i].prev, dots[i].pos)
+				dc.Stroke()
 			}
 		}
 	}
