@@ -1,6 +1,7 @@
 package main
 
 import (
+	"fmt"
 	"image"
 	"image/color"
 	"log"
@@ -9,6 +10,8 @@ import (
 	"net/http"
 	"strconv"
 	"time"
+
+	"github.com/schollz/progressbar/v3"
 
 	"github.com/rprtr258/gena"
 )
@@ -22,31 +25,34 @@ func cmap(r, m1, m2 float64) color.RGBA {
 	}
 }
 
+func newCanvas(w, h int) *image.RGBA {
+	return image.NewRGBA(image.Rect(0, 0, w, h))
+}
+
 func dotline() {
-	c := gena.NewCanvas(2080, 2080)
-	c.FillBackground(color.RGBA{230, 230, 230, 255})
+	c := newCanvas(2080, 2080)
+	gena.FillBackground(c, color.RGBA{230, 230, 230, 255})
 	gena.DotLine(c, gena.DarkPink, 10, 100, 20, 50, false, 15000)
-	gena.ToPNG(c.Img(), "dotline.png")
+	gena.SavePNG("dotline.png", c)
 }
 
 func dotswave() {
-	colors := []color.RGBA{
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Black)
+	gena.DotsWave(c, []color.RGBA{
 		{0xFF, 0xBE, 0x0B, 0xFF},
 		{0xFB, 0x56, 0x07, 0xFF},
 		{0xFF, 0x00, 0x6E, 0xFF},
 		{0x83, 0x38, 0xEC, 0xFF},
 		{0x3A, 0x86, 0xFF, 0xFF},
-	}
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Black)
-	gena.DotsWave(c, colors, 300)
-	gena.ToPNG(c.Img(), "dotswave.png")
+	}, 300)
+	gena.SavePNG("dotswave.png", c)
 }
 
 func gridsquare() {
-	c := gena.NewCanvas(600, 600)
+	c := newCanvas(600, 600)
 	gena.GirdSquares(c, gena.DarkPink, 24, 10, 0.2, 20)
-	gena.ToPNG(c.Img(), "gsquare.png")
+	gena.SavePNG("gsquare.png", c)
 }
 
 func _http() {
@@ -61,12 +67,12 @@ func _http() {
 			// Draw the image to bytes
 			b := func() []byte {
 				// Generate a new image
-				c := gena.NewCanvas(500, 500)
-				c.FillBackground(gena.Black)
+				c := newCanvas(500, 500)
+				gena.FillBackground(c, gena.Black)
 				gena.Janus(c, gena.DarkRed, gena.LightPink, 10, 0.2)
 
 				// Return the image as []byte
-				return gena.ToBytes(c.Img())
+				return gena.ToBytes(c)
 			}()
 
 			// Set content headers
@@ -81,40 +87,39 @@ func _http() {
 }
 
 func janus() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Black)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Black)
 	gena.Janus(c, gena.DarkRed, gena.LightPink, 10, 0.2)
-	gena.ToPNG(c.Img(), "janus.png")
+	gena.SavePNG("janus.png", c)
 }
 
 func julia() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Azure)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Azure)
 	gena.Julia(c, gena.Viridis, func(z complex128) complex128 {
 		return z*z + complex(-0.1, 0.651)
 	}, 40, 1.5+1.5i, 800)
-	gena.ToPNG(c.Img(), "julia.png")
+	gena.SavePNG("julia.png", c)
 }
 
 func maze() {
-	c := gena.NewCanvas(600, 600)
-	c.FillBackground(gena.Azure)
+	c := newCanvas(600, 600)
+	gena.FillBackground(c, gena.Azure)
 	gena.Maze(c, 3, gena.Orange, 20)
-	gena.ToPNG(c.Img(), "maze.png")
+	gena.SavePNG("maze.png", c)
 }
 
 func noiseline() {
-	colors := []color.RGBA{
+	c := newCanvas(1000, 1000)
+	gena.FillBackground(c, color.RGBA{0xF0, 0xFE, 0xFF, 0xFF})
+	gena.NoiseLine(c, []color.RGBA{
 		{0x06, 0x7B, 0xC2, 0xFF},
 		{0x84, 0xBC, 0xDA, 0xFF},
 		{0xEC, 0xC3, 0x0B, 0xFF},
 		{0xF3, 0x77, 0x48, 0xFF},
 		{0xD5, 0x60, 0x62, 0xFF},
-	}
-	c := gena.NewCanvas(1000, 1000)
-	c.FillBackground(color.RGBA{0xF0, 0xFE, 0xFF, 0xFF})
-	gena.NoiseLine(c, colors, 1000)
-	gena.ToPNG(c.Img(), "noiseline.png")
+	}, 1000)
+	gena.SavePNG("noiseline.png", c)
 }
 
 func oceanfish() {
@@ -132,50 +137,48 @@ func oceanfish() {
 		{0x19, 0x6E, 0x94, 0xFF},
 		{0x35, 0x3A, 0x57, 0xFF},
 	}
-	c := gena.NewCanvas(500, 500)
+	c := newCanvas(500, 500)
 	gena.OceanFish(c, colors, 100, 8)
-	gena.ToPNG(c.Img(), "oceanfish.png")
+	gena.SavePNG("oceanfish.png", c)
 }
 
 func perlinpearls() {
-	c := gena.
-		NewCanvas(500, 500)
-	c.FillBackground(gena.White)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.White)
 	gena.PerlinPearls(c, 0.3, 120, 10, 200, 40, 80, 200)
-	gena.ToPNG(c.Img(), "perlinperls.png")
+	gena.SavePNG("perlinperls.png", c)
 }
 
 func pixelhole() {
-	colors := []color.RGBA{
+	c := newCanvas(800, 800)
+	gena.FillBackground(c, gena.Black)
+	gena.PixelHole(c, []color.RGBA{
 		{0xF9, 0xC8, 0x0E, 0xFF},
 		{0xF8, 0x66, 0x24, 0xFF},
 		{0xEA, 0x35, 0x46, 0xFF},
 		{0x66, 0x2E, 0x9B, 0xFF},
 		{0x43, 0xBC, 0xCD, 0xFF},
-	}
-	c := gena.NewCanvas(800, 800)
-	c.FillBackground(gena.Black)
-	gena.PixelHole(c, colors, 60, 1200)
-	gena.ToPNG(c.Img(), "pixelhole.png")
+	}, 60, 1200)
+	gena.SavePNG("pixelhole.png", c)
 }
 
 func pointribbon() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Lavender)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Lavender)
 	gena.PointRibbon(c, 2, 50, 150000)
-	gena.ToPNG(c.Img(), "pointribbon.png")
+	gena.SavePNG("pointribbon.png", c)
 }
 
 func randcircle() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.MistyRose)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.MistyRose)
 	gena.RandCircle(c, gena.Plasma, 1, color.RGBA{122, 122, 122, 30}, 30, 80, 0.2, 2, 10, 30, true, 4)
-	gena.ToPNG(c.Img(), "randcircle.png")
+	gena.SavePNG("randcircle.png", c)
 }
 
 func randomshape() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.White)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.White)
 	gena.RandomShape(c, []color.RGBA{
 		{0xCF, 0x2B, 0x34, 0xFF},
 		{0xF0, 0x8F, 0x46, 0xFF},
@@ -183,137 +186,156 @@ func randomshape() {
 		{0x19, 0x6E, 0x94, 0xFF},
 		{0x35, 0x3A, 0x57, 0xFF},
 	}, 150)
-	gena.ToPNG(c.Img(), "randomshape.png")
+	gena.SavePNG("randomshape.png", c)
 }
 
 func silksky() {
-	c := gena.NewCanvas(600, 600)
+	c := newCanvas(600, 600)
 	gena.SilkSky(c, 10, 15, 5)
-	gena.ToPNG(c.Img(), "silksky.png")
+	gena.SavePNG("silksky.png", c)
 }
 
 func silksmoke() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Black)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Black)
 	gena.SilkSmoke(c, gena.Plasma, 1, gena.MediumAquamarine, 30, 400, 20, 0.2, 2, 10, 30, false)
-	gena.ToPNG(c.Img(), "silksmoke.png")
+	gena.SavePNG("silksmoke.png", c)
 }
 
 func solarflare() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Black)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Black)
 	gena.SolarFlare(c, color.RGBA{255, 64, 8, 128})
-	gena.ToPNG(c.Img(), "solarflare.png")
+	gena.SavePNG("solarflare.png", c)
 }
 
 func spiralsquare() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.MistyRose)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.MistyRose)
 	gena.SpiralSquare(c, gena.Plasma, 10, gena.Orange, 40, 400, 0.05, gena.Tomato, true)
-	gena.ToPNG(c.Img(), "spiralsquare.png")
+	gena.SavePNG("spiralsquare.png", c)
 }
 
 func swirl() {
 	bg := gena.Azure
-	c := gena.NewCanvas(1600, 1600)
-	c.FillBackground(bg)
-	gena.Swirl(c,
+	c := newCanvas(1600, 1600)
+	gena.FillBackground(c, bg)
+	gena.Swirl(
+		c,
 		color.RGBA{113, 3, 0, 140}, bg,
 		0.970, -1.899, 1.381, -1.506,
 		2.4+2.4i,
 		4000000,
 	)
-	gena.ToPNG(c.Img(), "swirl.png")
+	gena.SavePNG("swirl.png", c)
 }
 
 func yarn() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Orange)
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Orange)
 	gena.Yarn(c, 0.3, color.RGBA{A: 60}, 2000)
-	gena.ToPNG(c.Img(), "yarn.png")
+	gena.SavePNG("yarn.png", c)
 }
 
 func blackhole() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(color.RGBA{30, 30, 30, 255})
-
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, color.RGBA{30, 30, 30, 255})
 	gena.BlackHole(c, 1, gena.Tomato, 200, 400, 0.03, gena.NewPerlinNoiseDeprecated(), gena.RandomFloat64(0, 1))
-	gena.ToPNG(c.Img(), "blackhole.png")
+	gena.SavePNG("blackhole.png", c)
 }
 
 func circlegrid() {
-	colors := []color.RGBA{
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, color.RGBA{0xDF, 0xEB, 0xF5, 0xFF})
+	gena.CircleGrid(c, []color.RGBA{
 		{0xED, 0x34, 0x41, 0xFF},
 		{0xFF, 0xD6, 0x30, 0xFF},
 		{0x32, 0x9F, 0xE3, 0xFF},
 		{0x15, 0x42, 0x96, 0xFF},
 		{0x00, 0x00, 0x00, 0xFF},
 		{0xFF, 0xFF, 0xFF, 0xFF},
-	}
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(color.RGBA{0xDF, 0xEB, 0xF5, 0xFF})
-	gena.CircleGrid(c, colors, 2, 4, 6)
-	gena.ToPNG(c.Img(), "circlegrid.png")
+	}, 2, 4, 6)
+	gena.SavePNG("circlegrid.png", c)
 }
 
 func circleline() {
-	c := gena.NewCanvas(600, 600)
-	c.FillBackground(gena.Tan)
+	c := newCanvas(600, 600)
+	gena.FillBackground(c, gena.Tan)
 	gena.CircleLine(c, 1, gena.LightPink, 0.02, 600, 1.5, complex(2, 2))
-	gena.ToPNG(c.Img(), "circleline.png")
+	gena.SavePNG("circleline.png", c)
 }
 
 func circleloop() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Black)
+	c := newCanvas(500, 500)
 	gena.CircleLoop(c, 1, gena.Orange, 30, 100, 1000)
-	gena.ToPNG(c.Img(), "circleloop.png")
+	gena.SavePNG("circleloop.png", c)
 }
 
 func circleloop2() {
-	colors := []color.RGBA{
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, color.RGBA{8, 10, 20, 255})
+	gena.CircleLoop2(c, []color.RGBA{
 		{0xF9, 0xC8, 0x0E, 0xFF},
 		{0xF8, 0x66, 0x24, 0xFF},
 		{0xEA, 0x35, 0x46, 0xFF},
 		{0x66, 0x2E, 0x9B, 0xFF},
 		{0x43, 0xBC, 0xCD, 0xFF},
-	}
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(color.RGBA{8, 10, 20, 255})
-	gena.CircleLoop2(c, colors, 7)
-	gena.ToPNG(c.Img(), "colorloop2.png")
+	}, 7)
+	gena.SavePNG("colorloop2.png", c)
 }
 
 func circlemove() {
-	c := gena.NewCanvas(1200, 500)
-	c.FillBackground(gena.White)
+	c := newCanvas(1200, 500)
+	gena.FillBackground(c, gena.White)
 	gena.CircleMove(c, 1000)
-	gena.ToPNG(c.Img(), "circlemove.png")
+	gena.SavePNG("circlemove.png", c)
 }
 
 func circlenoise() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.White)
-	gena.CircleNoise(c, 0.3, 80, 2000, 60, 80, 400)
-	gena.ToPNG(c.Img(), "circlenoise.png")
+	const _dots = 2000
+	const _iters = 1 // 000
+	noise := gena.NewPerlinNoiseDeprecated()
+	bar := progressbar.Default(_iters)
+	for i := range _iters {
+		_ = bar.Add(1)
+
+		ralpha := float64(i) * 2 * math.Pi / _iters
+
+		angles := make([]float64, _dots)
+		for j := range angles {
+			angles[j] = noise.NoiseV2(complex(float64(j), 0)+gena.Polar(1, ralpha)) + ralpha
+		}
+
+		peepo := [4096]float64{}
+		for j := range peepo {
+			peepo[j] = noise.NoiseV2(complex(float64(j), 1) + gena.Polar(1, ralpha))
+		}
+		noise2 := gena.NewPerlinNoise(peepo)
+
+		c := newCanvas(500, 500)
+		gena.FillBackground(c, gena.White)
+		gena.CircleNoise(c, 0.3, 80, angles, 60, 80, 400, noise2)
+		gena.SavePNG(fmt.Sprintf("circlenoise%04d.png", i), c)
+	}
 }
 
 func colorcanvas() {
-	colors := []color.RGBA{
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Black)
+	gena.ColorCanva(c, []color.RGBA{
 		{0xF9, 0xC8, 0x0E, 0xFF},
 		{0xF8, 0x66, 0x24, 0xFF},
 		{0xEA, 0x35, 0x46, 0xFF},
 		{0x66, 0x2E, 0x9B, 0xFF},
 		{0x43, 0xBC, 0xCD, 0xFF},
-	}
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Black)
-	gena.ColorCanva(c, colors, 8, 5)
-	gena.ToPNG(c.Img(), "colorcanva.png")
+	}, 8, 5)
+	gena.SavePNG("colorcanva.png", c)
 }
 
 func colorcircle() {
-	colors := []color.RGBA{
+	c := newCanvas(1000, 1000)
+	gena.FillBackground(c, gena.White)
+	gena.ColorCircle(c, []color.RGBA{
 		{0xFF, 0xC6, 0x18, 0xFF},
 		{0xF4, 0x25, 0x39, 0xFF},
 		{0x41, 0x78, 0xF4, 0xFF},
@@ -322,63 +344,58 @@ func colorcircle() {
 		{0x56, 0xAC, 0x51, 0xFF},
 		{0x98, 0x19, 0xFA, 0xFF},
 		{0xFF, 0xFF, 0xFF, 0xFF},
-	}
-	c := gena.NewCanvas(1000, 1000)
-	c.FillBackground(gena.White)
-	gena.ColorCircle(c, colors, 500)
-	gena.ToPNG(c.Img(), "colorcircle.png")
+	}, 500)
+	gena.SavePNG("colorcircle.png", c)
 }
 
 func colorcircle2() {
-	colors := []color.RGBA{
+	c := newCanvas(800, 800)
+	gena.FillBackground(c, gena.White)
+	gena.ColorCircle2(c, []color.RGBA{
 		{0x11, 0x60, 0xC6, 0xFF},
 		{0xFD, 0xD9, 0x00, 0xFF},
 		{0xF5, 0xB4, 0xF8, 0xFF},
 		{0xEF, 0x13, 0x55, 0xFF},
 		{0xF4, 0x9F, 0x0A, 0xFF},
-	}
-	c := gena.NewCanvas(800, 800)
-	c.FillBackground(gena.White)
-	gena.ColorCircle2(c, colors, 30)
-	gena.ToPNG(c.Img(), "colorcircle2.png")
+	}, 30)
+	gena.SavePNG("colorcircle2.png", c)
 }
 
 func contourline() {
-	colors := []color.RGBA{
+	c := newCanvas(1600, 1600)
+	gena.FillBackground(c, color.RGBA{0x1a, 0x06, 0x33, 0xFF})
+	gena.ContourLine(c, []color.RGBA{
 		{0x58, 0x18, 0x45, 0xFF},
 		{0x90, 0x0C, 0x3F, 0xFF},
 		{0xC7, 0x00, 0x39, 0xFF},
 		{0xFF, 0x57, 0x33, 0xFF},
 		{0xFF, 0xC3, 0x0F, 0xFF},
-	}
-	c := gena.NewCanvas(1600, 1600)
-	c.FillBackground(color.RGBA{0x1a, 0x06, 0x33, 0xFF})
-	gena.ContourLine(c, colors, 500)
-	gena.ToPNG(c.Img(), "contourline.png")
+	}, 500)
+	gena.SavePNG("contourline.png", c)
 }
 
 func domainwrap() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Black)
-	gena.DomainWrap(c, 0.01, 4, complex(4, 20), cmap, 0, 0, "")
-	gena.ToPNG(c.Img(), "domainwarp.png")
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Black)
+	gena.DomainWarp(c, 0.01, 4, complex(4, 20), cmap, 0, 0, "")
+	gena.SavePNG("domainwarp.png", c)
 }
 
 func domainwrapFrames() {
-	c := gena.NewCanvas(500, 500)
-	c.FillBackground(gena.Black)
-	gena.DomainWrap(c, 0.01, 4, complex(4, 20), cmap, 0.005, 100, "./temp")
+	c := newCanvas(500, 500)
+	gena.FillBackground(c, gena.Black)
+	gena.DomainWarp(c, 0.01, 4, complex(4, 20), cmap, 0.005, 100, "./temp")
 }
 
 func test() {
 	dest := image.NewRGBA(image.Rect(0, 0, 500, 500))
 
-	ctex := gena.NewContextForRGBA(dest)
-	ctex.Stack(func(ctx *gena.Context) {
-		ctex.Translate(complex(500/2, 500/2))
-		ctex.Rotate(40)
-		ctex.SetColor(color.RGBA{0xFF, 0x00, 0x00, 255})
-		for i := 0; i < 361; i++ {
+	dc := gena.NewContextForRGBA(dest)
+	dc.Stack(func(dc *gena.Context) {
+		dc.Translate(complex(500/2, 500/2))
+		dc.Rotate(40)
+		dc.SetColor(color.RGBA{0xFF, 0x00, 0x00, 255})
+		for i := range 361 {
 			theta := gena.Radians(float64(i))
 			p := complex(
 				math.Cos(theta)-math.Pow(math.Sin(theta), 2)/math.Sqrt(2),
@@ -392,11 +409,11 @@ func test() {
 				math.Cos(alpha)*math.Sin(alpha),
 			)
 
-			ctex.DrawLine(p*100, p1*100)
-			ctex.Stroke()
+			dc.DrawLine(p*100, p1*100)
+			dc.Stroke()
 		}
 	})
-	ctex.SavePNG("test.png")
+	gena.SavePNG("test.png", dc.Image())
 }
 
 func out() {
@@ -408,7 +425,7 @@ func out() {
 	dc.DrawRectangle(0, complex(1000, 1000))
 	dc.SetRGB(0, 0, 0)
 	dc.Fill()
-	dc.SavePNG("out.png")
+	gena.SavePNG("out.png", dc.Image())
 }
 
 func main() {
