@@ -11,13 +11,13 @@ func unicode() {
 	dc.Clear()
 	dc.SetRGB(0, 0, 0)
 	dc.LoadFontFace("Xolonium-Regular.ttf", F)
-	for r := 0; r < 256; r++ {
-		for c := 0; c < 256; c++ {
+	for r := range 256 {
+		for c := range 256 {
 			i := r*256 + c
 			x := float64(c*T) + T/2
 			y := float64(r*T) + T/2
 			dc.DrawStringAnchored(string(rune(i)), x, y, 0.5, 0.5)
 		}
 	}
-	dc.SavePNG("unicode.png")
+	gena.SavePNG("unicode.png", dc.Image())
 }
