@@ -1,24 +1,25 @@
 package gena
 
 import (
+	"image"
 	"image/color"
+	"math/cmplx"
 	"math/rand"
 )
 
-type colorCanva struct {
-	seg float64
+type Rect struct {
+	Pos, Size V2
 }
 
 // ColorCanva returns a color canva image.
-func ColorCanva(c Canvas, colorSchema []color.RGBA, lineWidth float64, seg float64) {
-	cc := &colorCanva{seg: seg}
-	dc := NewContextForRGBA(c.Img())
+func ColorCanva(c *image.RGBA, colorSchema []color.RGBA, lineWidth float64, seg float64) {
+	dc := NewContextForRGBA(c)
 	dc.SetLineWidth(lineWidth)
 
 	rects := make([]Rect, 0)
-	w := float64(c.Width) / cc.seg
-	for i := 0.; i < cc.seg; i += 1. {
-		for j := 0.; j < cc.seg; j += 1. {
+	w := float64(c.Bounds().Dx()) / seg
+	for i := 0.; i < seg; i += 1. {
+		for j := 0.; j < seg; j += 1. {
 			rects = append(rects, Rect{
 				Pos:  Plus(Mul(complex(i, j), w), w/2),
 				Size: complex(w, w),
@@ -30,28 +31,32 @@ func ColorCanva(c Canvas, colorSchema []color.RGBA, lineWidth float64, seg float
 		rects[i], rects[j] = rects[j], rects[i]
 	})
 
-	dc.Translate(c.Size() / 2)
+	dc.Translate(Size(c) / 2)
 	dc.Scale(complex(0.6, 0.6))
-	dc.Translate(-c.Size() / 2)
+	dc.Translate(-Size(c) / 2)
 
-	for i := 0; i < len(rects); i++ {
-		cc.draw(c, dc, colorSchema, rects[i])
-		cc.draw(c, dc, colorSchema, rects[i])
+	for i := range len(rects) {
+		drawColorCanva(dc, seg, colorSchema, rects[i])
+		drawColorCanva(dc, seg, colorSchema, rects[i])
 	}
 }
 
-func (cc *colorCanva) draw(c Canvas, dc *Context, colorSchema []color.RGBA, rect Rect) {
-	wh := Mul2(rect.Size/5, Mul(RandomV2(), cc.seg*2)+1)
+func drawColorCanva(dc *Context, seg float64, colorSchema []color.RGBA, rect Rect) {
+	wh := Mul2(rect.Size/5, Mul(RandomV2(), seg*2)+1)
+
+	var delta V2
 	switch rand.Intn(4) {
 	case 0:
-		dc.DrawRectangle(complex(X(rect.Pos)-X(wh)/2+X(rect.Size)/2, Y(rect.Pos)+Y(wh)/2+Y(rect.Size)/2), wh)
+		delta = complex(-X(rect.Size), -Y(rect.Size))
 	case 1:
-		dc.DrawRectangle(complex(X(rect.Pos)-X(wh)/2-X(rect.Size)/2, Y(rect.Pos)+Y(wh)/2+Y(rect.Size)/2), wh)
+		delta = complex(X(rect.Size), -Y(rect.Size))
 	case 2:
-		dc.DrawRectangle(complex(X(rect.Pos)-X(wh)/2+X(rect.Size)/2, Y(rect.Pos)+Y(wh)/2-Y(rect.Size)/2), wh)
+		delta = complex(-X(rect.Size), Y(rect.Size))
 	case 3:
-		dc.DrawRectangle(complex(X(rect.Pos)-X(wh)/2-X(rect.Size)/2, Y(rect.Pos)+Y(wh)/2-Y(rect.Size)/2), wh)
+		delta = complex(X(rect.Size), Y(rect.Size))
 	}
+	dc.DrawRectangle(rect.Pos-cmplx.Conj(wh)/2+delta/2, wh)
+
 	dc.SetColor(Black)
 	dc.StrokePreserve()
 	dc.SetColor(colorSchema[rand.Intn(len(colorSchema))])
