@@ -1,6 +1,7 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math/rand"
 )
@@ -8,31 +9,30 @@ import (
 // DotLine would draw images with the short dot and short.
 // The short lines would compose as a big circle.
 //   - n: The number of elements in this image.
-//   - ras/canv: Control the appearance of this image.
+//   - ras, canv: Control the appearance of this image.
 //   - randColor: Use the specified color or random colors.
-func DotLine(c Canvas, colorSchema []color.RGBA, lineWidth float64, n int, ras, canv float64, randColor bool, iters int) {
-	dc := NewContextForRGBA(c.Img())
-
+func DotLine(c *image.RGBA, colorSchema []color.RGBA, lineWidth float64, n int, ras, canv float64, randColor bool, iters int) {
+	dc := NewContextForRGBA(c)
 	dc.SetLineWidth(lineWidth)
-	var dir []int = []int{-1, 1}
-	for i := 0; i < iters; i++ {
+	dir := []int{-1, 1}
+	for range iters {
 		old := complex(
 			float64(rand.Intn(n-1)),
 			float64(rand.Intn(n-1)),
 		)
 
-		n := rand.Intn(7)
+		k := rand.Intn(7)
 		if randColor {
 			dc.SetColor(colorSchema[rand.Intn(len(colorSchema))])
 		} else {
 			dc.SetRGBA255(color.RGBA{
-				uint8(RandomRangeInt(222, 255)),
-				uint8(RandomRangeInt(20, 222)),
+				RandomRangeInt[uint8](222, 255),
+				RandomRangeInt[uint8](20, 222),
 				0,
 				0,
 			}, 255)
 		}
-		for j := 0; j < n; j++ {
+		for range k {
 			new := old + complex(float64(dir[rand.Intn(2)]), float64(dir[rand.Intn(2)]))
 			if Dist(new, complex(float64(n/2), float64(n/2))) > float64(n/2-10) {
 				new = old
@@ -42,12 +42,8 @@ func DotLine(c Canvas, colorSchema []color.RGBA, lineWidth float64, n int, ras, 
 				dc.Fill()
 				continue
 			}
-			dc.DrawLine(
-				Plus(Mul(old, ras), canv),
-				Plus(Mul(new, ras), canv),
-			)
+			dc.DrawLine(Plus(Mul(old, ras), canv), Plus(Mul(new, ras), canv))
 			old = new
-
 			dc.Stroke()
 		}
 	}
