@@ -1,6 +1,7 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math"
 	"math/rand"
@@ -14,24 +15,24 @@ type oceanFish struct {
 // OceanFish draws an ocean and some fishes in the center.
 //   - lineNum: The number of the line used to simulate the ocean wave.
 //   - fishNum: The number of fish.
-func OceanFish(c Canvas, colorSchema []color.RGBA, lineNum, fishNum int) {
+func OceanFish(c *image.RGBA, colorSchema []color.RGBA, lineNum, fishNum int) {
 	o := &oceanFish{
 		lineNum: lineNum,
 		fishNum: fishNum,
 	}
 
-	dc := NewContextForRGBA(c.Img())
+	dc := NewContextForRGBA(c)
 
 	o.drawlines(dc, c, colorSchema)
 
-	for i := 0; i < o.fishNum; i++ {
+	for i := range o.fishNum {
 		dc.Stack(func(ctx *Context) {
 			dc.Stack(func(ctx *Context) {
 				theta := float64(360*i) / float64(o.fishNum)
-				r := float64(c.Width) / 4.0
-				dc.Translate(Mul2(c.Size()/2, Polar(r, Radians(theta))))
+				r := float64(c.Bounds().Dx()) / 4.0
+				dc.Translate(Mul2(Size(c)/2, Polar(r, Radians(theta))))
 				dc.Rotate(Radians(theta + 90))
-				o.drawfish(dc, c, 0, float64(c.Width)/10)
+				o.drawfish(dc, c, 0, float64(c.Bounds().Dx())/10)
 			})
 			dc.Clip()
 			o.drawlines(dc, c, colorSchema)
@@ -41,15 +42,15 @@ func OceanFish(c Canvas, colorSchema []color.RGBA, lineNum, fishNum int) {
 	}
 }
 
-func (o *oceanFish) drawlines(ctx *Context, c Canvas, colorSchema []color.RGBA) {
-	for i := 0; i < o.lineNum; i++ {
+func (o *oceanFish) drawlines(ctx *Context, c *image.RGBA, colorSchema []color.RGBA) {
+	for range o.lineNum {
 		cl := colorSchema[rand.Intn(len(colorSchema))]
 		ctx.SetColor(cl)
 		ctx.SetLineWidth(RandomFloat64(3, 20))
-		y := rand.Float64() * float64(c.Height)
+		y := rand.Float64() * float64(c.Bounds().Dy())
 		ctx.DrawLine(
 			complex(0, y+RandomFloat64(-50, 50)),
-			complex(float64(c.Width), y+RandomFloat64(-50, 50)),
+			complex(float64(c.Bounds().Dx()), y+RandomFloat64(-50, 50)),
 		)
 		ctx.Stroke()
 	}
@@ -62,7 +63,7 @@ func fishPt(r, theta float64) V2 {
 	)
 }
 
-func (o *oceanFish) drawfish(dc *Context, c Canvas, v V2, r float64) {
+func (o *oceanFish) drawfish(dc *Context, c *image.RGBA, v V2, r float64) {
 	dc.Stack(func(ctx *Context) {
 		dc.Translate(v)
 		dc.Rotate(Radians(180))
