@@ -8,7 +8,7 @@ import (
 )
 
 // SolarFlare draws a solar flare images.
-func SolarFlare(c Canvas, lineColor color.RGBA) {
+func SolarFlare(c *image.RGBA, lineColor color.RGBA) {
 	var xOffset, yOffset float64
 	const offsetInc = 0.006
 	const inc = 1.0
@@ -16,16 +16,16 @@ func SolarFlare(c Canvas, lineColor color.RGBA) {
 	noise := NewPerlinNoiseDeprecated()
 
 	for r := 1.0; r < 200; {
-		for i := 0; i < 10; i++ {
+		for range 10 {
 			nPoints := int(2 * math.Pi * r)
 			nPoints = min(nPoints, 500)
 
-			img := image.NewRGBA(image.Rect(0, 0, c.Width, c.Height))
+			img := image.NewRGBA(image.Rect(0, 0, c.Bounds().Dx(), c.Bounds().Dy()))
 			draw.Draw(img, img.Bounds(), &image.Uniform{color.Black}, image.Point{}, draw.Src)
 			dc := NewContextForRGBA(img)
 
 			dc.Stack(func(ctx *Context) {
-				dc.Translate(c.Size() / 2)
+				dc.Translate(Size(c) / 2)
 				dc.SetLineWidth(1.0)
 				dc.SetColor(lineColor)
 				for j := 0; j < nPoints+1; j += 1 {
@@ -40,7 +40,7 @@ func SolarFlare(c Canvas, lineColor color.RGBA) {
 				dc.Stroke()
 			})
 
-			c.img = Blend(img, c.img, Add)
+			c = Blend(img, c, Add)
 			xOffset += offsetInc
 			yOffset += offsetInc
 			r *= m
