@@ -1,26 +1,27 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math/rand"
 )
 
 // SpiralSquare draws a spiral square images.
 func SpiralSquare(
-	c Canvas, colorSchema []color.RGBA,
+	c *image.RGBA, colorSchema []color.RGBA,
 	lineWidth float64,
 	lineColor color.RGBA,
 	squareNum int,
 	rectSide, decay float64,
 	fg color.RGBA, randColor bool,
 ) {
-	dc := NewContextForRGBA(c.Img())
+	dc := NewContextForRGBA(c)
 
 	sl := rectSide
 	theta := rand.Intn(360) + 1
-	for i := 0; i < squareNum; i++ {
+	for i := range squareNum {
 		dc.Stack(func(ctx *Context) {
-			dc.Translate(c.Size() / 2)
+			dc.Translate(Size(c) / 2)
 			dc.Rotate(Radians(float64(theta * (i + 1))))
 
 			dc.Scale(complex(sl, sl))
