@@ -1,6 +1,7 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math"
 )
@@ -21,53 +22,53 @@ func (s *swirl) swirlTransform(p V2) V2 {
 	)
 }
 
-func (s *swirl) removeNoisy(c Canvas) {
-	for i := 1; i < c.Width-1; i++ {
-		for j := 1; j < c.Height-1; j++ {
-			if c.Img().At(i, j) == s.bg {
+func (s *swirl) removeNoisy(c *image.RGBA) {
+	for i := 1; i < c.Bounds().Dx()-1; i++ {
+		for j := 1; j < c.Bounds().Dy()-1; j++ {
+			if c.At(i, j) == s.bg {
 				continue
 			}
 
 			var n int
-			if c.Img().At(i+1, j) == s.bg {
+			if c.At(i+1, j) == s.bg {
 				n += 1
 			}
-			if c.Img().At(i+1, j+1) == s.bg {
+			if c.At(i+1, j+1) == s.bg {
 				n += 1
 			}
-			if c.Img().At(i, j+1) == s.bg {
-				n += 1
-			}
-
-			if c.Img().At(i-1, j) == s.bg {
+			if c.At(i, j+1) == s.bg {
 				n += 1
 			}
 
-			if c.Img().At(i-1, j+1) == s.bg {
+			if c.At(i-1, j) == s.bg {
 				n += 1
 			}
 
-			if c.Img().At(i-1, j-1) == s.bg {
+			if c.At(i-1, j+1) == s.bg {
 				n += 1
 			}
 
-			if c.Img().At(i+1, j-1) == s.bg {
+			if c.At(i-1, j-1) == s.bg {
 				n += 1
 			}
 
-			if c.Img().At(i, j-1) == s.bg {
+			if c.At(i+1, j-1) == s.bg {
+				n += 1
+			}
+
+			if c.At(i, j-1) == s.bg {
 				n += 1
 			}
 
 			if n > 5 {
-				c.Img().Set(i, j, s.bg)
+				c.Set(i, j, s.bg)
 			}
 		}
 	}
 }
 
 // Swirl draws a swirl image.
-func Swirl(C Canvas, fg, bg color.RGBA, a, b, c, d float64, axis V2, iters int) {
+func Swirl(img *image.RGBA, fg, bg color.RGBA, a, b, c, d float64, axis V2, iters int) {
 	s := &swirl{
 		fg:    fg,
 		bg:    bg,
@@ -78,17 +79,17 @@ func Swirl(C Canvas, fg, bg color.RGBA, a, b, c, d float64, axis V2, iters int) 
 		axis:  axis,
 		iters: iters,
 	}
-	ctex := NewContextForRGBA(C.Img())
-	ctex.SetLineWidth(3)
+	dc := NewContextForRGBA(img)
+	dc.SetLineWidth(3)
 	start := complex(1.0, 1.0)
 
-	for i := 0; i < s.iters; i++ {
+	for range s.iters {
 		next := s.swirlTransform(start)
-		xy := ToPixel(next, s.axis, C.Size())
-		C.Img().Set(int(X(xy)), int(Y(xy)), s.fg)
+		xy := ToPixel(next, s.axis, Size(img))
+		img.Set(int(X(xy)), int(Y(xy)), s.fg)
 		start = next
 	}
 
-	s.removeNoisy(C)
-	s.removeNoisy(C)
+	s.removeNoisy(img)
+	s.removeNoisy(img)
 }
