@@ -1,6 +1,7 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math/cmplx"
 )
@@ -13,14 +14,14 @@ type GenFunc func(complex128) complex128
 //   - fn: The custom julia set function.
 //   - maxz: The maximum modulus length of a complex number.
 //   - axis: The range for the X-Y coordination used to mapping the julia set number to the real pixel of the image. These should be positive. It only indicates the first quadrant range.
-func Julia(c Canvas, colorSchema []color.RGBA, fn GenFunc, maxz float64, axis V2, iters int) {
+func Julia(c *image.RGBA, colorSchema []color.RGBA, fn GenFunc, maxz float64, axis V2, iters int) {
 	n := uint8(min(len(colorSchema), 255))
 
-	for i := 0; i <= c.Width; i++ {
-		for k := 0; k <= c.Height; k++ {
+	for i := 0; i <= c.Bounds().Dx(); i++ {
+		for k := 0; k <= c.Bounds().Dy(); k++ {
 			z := complex(
-				float64(i)/float64(c.Width)*2.0*Y(axis)-Y(axis),
-				float64(k)/float64(c.Height)*2.0*Y(axis)-Y(axis),
+				float64(i)/float64(c.Bounds().Dx())*2.0*Y(axis)-Y(axis),
+				float64(k)/float64(c.Bounds().Dy())*2.0*Y(axis)-Y(axis),
 			)
 
 			var nit int
@@ -28,7 +29,7 @@ func Julia(c Canvas, colorSchema []color.RGBA, fn GenFunc, maxz float64, axis V2
 				z = fn(z)
 			}
 
-			c.Img().Set(i, k, colorSchema[uint8(nit*255/iters)%n])
+			c.Set(i, k, colorSchema[uint8(nit*255/iters)%n])
 		}
 	}
 }
