@@ -1,30 +1,31 @@
 package gena
 
 import (
+	"image"
 	"image/color"
 	"math/rand"
 )
 
 // Maze draws a random maze image.
-func Maze(c Canvas, lineWidth float64, lineColor color.RGBA, step int) {
-	ctex := NewContextForRGBA(c.Img())
-	ctex.SetColor(lineColor)
-	ctex.SetLineWidth(lineWidth)
-	for x := 0; x < c.Width; x += step {
-		for y := 0; y < c.Height; y += step {
+func Maze(c *image.RGBA, lineWidth float64, lineColor color.RGBA, step int) {
+	dc := NewContextForRGBA(c)
+	dc.SetColor(lineColor)
+	dc.SetLineWidth(lineWidth)
+	for x := 0; x < c.Bounds().Dx(); x += step {
+		for y := 0; y < c.Bounds().Dy(); y += step {
 			v := complex(float64(x), float64(y))
 			if rand.Float32() > 0.5 {
-				ctex.DrawLine(
+				dc.DrawLine(
 					v,
 					v+complex(float64(step), float64(step)),
 				)
 			} else {
-				ctex.DrawLine(
+				dc.DrawLine(
 					v+complex(float64(step), 0),
 					v+complex(0, float64(step)),
 				)
 			}
-			ctex.Stroke()
+			dc.Stroke()
 		}
 	}
 }
