@@ -1,38 +1,29 @@
 package gena
 
-import "image/color"
+import (
+	"image"
+	"image/color"
+)
 
 // Yarn draws a yarn image.
 //   - n: The number of the curve.
-func Yarn(c Canvas, lineWidth float64, lineColor color.RGBA, n int) {
-	ctex := NewContextForRGBA(c.Img())
-	ctex.SetLineWidth(lineWidth)
-	ctex.SetColor(lineColor)
+func Yarn(c *image.RGBA, lineWidth float64, lineColor color.RGBA, n int) {
+	dc := NewContextForRGBA(c)
+	dc.SetLineWidth(lineWidth)
+	dc.SetColor(lineColor)
 	noise := NewPerlinNoiseDeprecated()
 
 	offset := 0.0
 	inc := 0.005
-	for i := 0; i < n; i++ {
-		p0 := Mul2(complex(
-			noise.Noise1D(offset+15),
-			noise.Noise1D(offset+55),
-		), c.Size())
-		p1 := Mul2(complex(
-			noise.Noise1D(offset+25),
-			noise.Noise1D(offset+65),
-		), c.Size())
-		p2 := Mul2(complex(
-			noise.Noise1D(offset+35),
-			noise.Noise1D(offset+75),
-		), c.Size())
-		p3 := Mul2(complex(
-			noise.Noise1D(offset+45),
-			noise.Noise1D(offset+85),
-		), c.Size())
-		ctex.MoveToV2(p0)
-		ctex.CubicTo(p1, p2, p3)
-		ctex.Stroke()
-		ctex.ClearPath()
+	for range n {
+		dc.MoveToV2(Mul2(noise.NoiseV2D2(offset+15, offset+55), Size(c)))
+		dc.CubicTo(
+			Mul2(noise.NoiseV2D2(offset+25, offset+65), Size(c)),
+			Mul2(noise.NoiseV2D2(offset+35, offset+75), Size(c)),
+			Mul2(noise.NoiseV2D2(offset+45, offset+85), Size(c)),
+		)
+		dc.Stroke()
+		dc.ClearPath()
 		offset += inc
 	}
 }
